pyproject.toml
[tool.poetry]
name = "hw-m6"
version = "0.1.0"
description = ""
authors = ["mx_so <108818234+m-rznv@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "hw_m6"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
